+++ crates/auto-lang/test/a2r/12_specs/010_spec_supertrait/spec_supertrait.expected.rs
@@ -11,5 +11,5 @@ trait Clone: Send {
 
 
 fn main() {
-    let x: i32 = 42;
+    let x: i64 = 42;
 }
