+++ crates/auto-lang/test/a2r/16_interop/001_async_fn/async_fn.expected.rs
@@ -5,5 +5,5 @@ async fn fetch_data(url: &str) -> String {
 }
 
 fn main() {
-    let x: i32 = 42;
+    let x: i64 = 42;
 }
