+++ crates/auto-lang/test/a2r/16_interop/017_comptime_read_text/comptime_read_text.expected.rs
@@ -4,7 +4,7 @@ fn banner() -> String {
     return "Hello from comptime!";
 }
 
-fn answer() -> i32 {
+fn answer() -> i64 {
     return 42;
 }
