+++ crates/auto-lang/test/a2r/19_ownership/008_indirect_self/indirect_self.expected.rs
@@ -2,7 +2,7 @@
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 struct Tree {
-    pub value: i32,
+    pub value: i64,
     pub children: Vec<Tree>,
 }
