+++ crates/auto-lang/test/cookbook/encoding/009_csv_invalid/csv_invalid.expected.rs
@@ -8,7 +8,7 @@ use a2r_std::*;
 fn main() {
     let data: String = "name,age\nAlice,30\nBob,twenty-five\nCharlie,35".to_string();
     let mut lines = data.split("\n").collect::<Vec<_>>();
-    assert!((lines.len() as i32) == 4);
+    assert!((lines.len() as i64) == 4);
 
     assert!(a2r_std::str_contains(data.as_str(), "twenty-five"));
 }
