+++ crates/auto-lang/test/cookbook/os/003_error_file/error_file.expected.rs
@@ -6,5 +6,5 @@ use std::process::Stdio;
 fn main() {
     let result = Command::new("echo").arg("test output").output();
     let stdout = result.stdout.clone();
-    assert!((stdout.len() as i32) > 0);
+    assert!((stdout.len() as i64) > 0);
 }
